Move MaybeProxyClient to edgelet-http
