+++ edgelet/edgelet-http/Cargo.toml
@@ -12,6 +12,8 @@ failure_derive = "0.1"
 futures = "0.1"
 http = "0.1"
 hyper = { version = "0.11", features = ["compat"] }
+hyper-proxy = "0.4.1"
+hyper-tls = "0.1"
 log = "0.4"
 percent-encoding = "1.0"
 regex = "0.2"
