+++ edgelet/iotedged/Cargo.toml
@@ -12,7 +12,6 @@ env_logger = "0.5"
 failure = "0.1"
 futures = "0.1"
 hyper = "0.11"
-hyper-proxy = "0.4.1"
 hyper-tls = "0.1"
 log = "0.4"
 serde = "1.0"
