+++ edgelet/iotedged/src/lib.rs
@@ -23,7 +23,6 @@ extern crate failure;
 extern crate futures;
 extern crate hsm;
 extern crate hyper;
-extern crate hyper_proxy;
 extern crate hyper_tls;
 extern crate iothubservice;
 #[macro_use]
@@ -78,7 +77,7 @@ use edgelet_hsm::tpm::{TpmKey, TpmKeyStore};
 use edgelet_hsm::Crypto;
 use edgelet_http::client::Client as HttpClient;
 use edgelet_http::logging::LoggingService;
-use edgelet_http::{ApiVersionService, HyperExt, API_VERSION};
+use edgelet_http::{ApiVersionService, HyperExt, MaybeProxyClient, API_VERSION};
 use edgelet_http_mgmt::ManagementService;
 use edgelet_http_workload::WorkloadService;
 use edgelet_iothub::{HubIdentityManager, SasTokenSource};
@@ -88,11 +87,9 @@ use futures::sync::oneshot::{self, Receiver};
 use futures::Future;
 use hsm::tpm::Tpm;
 use hsm::ManageTpmKeys;
-use hyper::client::{FutureResponse, HttpConnector, Service};
+use hyper::client::Service;
 use hyper::server::Http;
-use hyper::{Client as HyperClient, Error as HyperError, Request, Response, Uri};
-use hyper_proxy::{Proxy, ProxyConnector, Intercept};
-use hyper_tls::HttpsConnector;
+use hyper::{Error as HyperError, Request, Response, Uri};
 use iothubservice::DeviceClient;
 use provisioning::provisioning::{
     BackupProvisioning, DpsProvisioning, ManualProvisioning, Provision, ProvisioningResult,
@@ -164,7 +161,6 @@ const EDGE_NETWORKID_KEY: &str = "NetworkId";
 const API_VERSION_KEY: &str = "IOTEDGE_APIVERSION";
 
 const IOTHUB_API_VERSION: &str = "2017-11-08-preview";
-const DNS_WORKER_THREADS: usize = 4;
 const UNIX_SCHEME: &str = "unix";
 
 /// This is the name of the provisioning backup file
@@ -183,42 +179,6 @@ const IOTEDGED_COMMONNAME: &str = "iotedged workload ca";
 pub struct Main {
     settings: Settings<DockerConfig>,
     reactor: Core,
-}
-
-#[derive(Clone)]
-pub enum MaybeProxyClient {
-    NoProxy(HyperClient<HttpsConnector<HttpConnector>>),
-    Proxy(HyperClient<ProxyConnector<HttpsConnector<HttpConnector>>>),
-}
-
-impl MaybeProxyClient {
-    pub fn new(handle: &Handle, proxy_uri: Option<Uri>) -> Result<MaybeProxyClient, Error> {
-        let https = HttpsConnector::new(DNS_WORKER_THREADS, &handle.clone())?;
-        match proxy_uri {
-            None => {
-                Ok(MaybeProxyClient::NoProxy(HyperClient::configure().connector(https).build(&handle)))
-            },
-            Some(uri) => {
-                let proxy = Proxy::new(Intercept::All, uri);
-                let connector = ProxyConnector::from_proxy(https, proxy)?;
-                Ok(MaybeProxyClient::Proxy(HyperClient::configure().connector(connector).build(&handle)))
-            }
-        }
-    }
-}
-
-impl Service for MaybeProxyClient {
-    type Request = Request;
-    type Response = Response;
-    type Error = HyperError;
-    type Future = FutureResponse;
-
-    fn call(&self, req: Self::Request) -> Self::Future {
-        match *self {
-            MaybeProxyClient::NoProxy(ref client) => client.call(req) as Self::Future,
-            MaybeProxyClient::Proxy(ref client) => client.call(req) as Self::Future,
-        }
-    }
 }
 
 impl Main {
